pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.35.3",
    "pandas>=2.3.3",
    "datasets>=2.0.0",
    "requests>=2.25.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
